Main functionality is ready

+++ server/controllers/newsCtrl.js
@@ -9,22 +9,22 @@ const News = mongoose.model('News');
 
 module.exports.getAllNews = async (req, res) => {
   try {
-    const accessToken = req.headers['authorization'];
-    const result = await jwt.verify(accessToken, secret);
-
-    if (result) {
-      const foundedNews = await News.find();
-      if (!foundedNews) {
-        throw new Error(`Неправильный access токен!`);
-      }
-
-      const preparedAllNews = [];
-      foundedNews.map((news) =>
-        preparedAllNews.push(serialize.serializeNews(news))
-      );
-
-      res.json(preparedAllNews);
+    const foundedUser = await User.findOne({
+      _id: req.accessTokenData.payload,
+    });
+    if (!foundedUser.permission.news.R) {
+      throw new Error(`Нет прав на чтение новостей!`);
     }
+
+    const foundedAllNews = await News.find();
+    if (!foundedAllNews) {
+      res.json([]);
+    }
+
+    const preparedAllNews = foundedAllNews.map((news) => {
+      return serialize.serializeNews(news);
+    });
+    res.json(preparedAllNews);
   } catch (err) {
     res.status(401).json({ message: err.message });
   }
@@ -32,12 +32,11 @@ module.exports.getAllNews = async (req, res) => {
 
 module.exports.createNews = async (req, res) => {
   try {
-    const accessToken = req.headers['authorization'];
-    const result = await jwt.verify(accessToken, secret);
-
-    const foundedUser = await User.findOne({ _id: result.payload });
-    if (!foundedUser) {
-      throw new Error(`Пользователь ${username} не зарегистрирован!`);
+    const foundedUser = await User.findOne({
+      _id: req.accessTokenData.payload,
+    });
+    if (!foundedUser.permission.news.C) {
+      throw new Error(`Нет прав на создание новостей!`);
     }
 
     const preparedForCreate = {
@@ -55,16 +54,14 @@ module.exports.createNews = async (req, res) => {
     };
 
     const createdNews = await News.create(preparedForCreate);
-
-    const foundedNews = await News.find();
-    if (!foundedNews) {
-      throw new Error(`Неправильный access токен!`);
+    const foundedAllNews = await News.find();
+    if (!foundedAllNews) {
+      res.json([]);
     }
 
-    const preparedAllNews = [];
-    foundedNews.map((news) =>
-      preparedAllNews.push(serialize.serializeNews(news))
-    );
+    const preparedAllNews = foundedAllNews.map((news) => {
+      return serialize.serializeNews(news);
+    });
 
     res.json(preparedAllNews);
   } catch (err) {
@@ -74,8 +71,12 @@ module.exports.createNews = async (req, res) => {
 
 module.exports.changeNews = async (req, res) => {
   try {
-    const accessToken = req.headers['authorization'];
-    const result = await jwt.verify(accessToken, secret);
+    const foundedUser = await User.findOne({
+      _id: req.accessTokenData.payload,
+    });
+    if (!foundedUser.permission.news.U) {
+      throw new Error(`Нет прав на обновление новостей!`);
+    }
 
     const foundedNews = await News.findOneAndUpdate(
       { _id: req.params.id },
@@ -83,18 +84,17 @@ module.exports.changeNews = async (req, res) => {
       { new: true }
     );
     if (!foundedNews) {
-      throw new Error(`Пользователь ${username} не зарегистрирован!`);
+      throw new Error(`Ошибка при обновлении новости!`);
     }
 
     const foundedAllNews = await News.find();
     if (!foundedAllNews) {
-      throw new Error(`Неправильный access токен!`);
+      res.json([]);
     }
 
-    const preparedAllNews = [];
-    foundedAllNews.map((news) =>
-      preparedAllNews.push(serialize.serializeNews(news))
-    );
+    const preparedAllNews = foundedAllNews.map((news) => {
+      return serialize.serializeNews(news);
+    });
 
     res.json(preparedAllNews);
   } catch (err) {
@@ -104,8 +104,12 @@ module.exports.changeNews = async (req, res) => {
 
 module.exports.deleteNews = async (req, res) => {
   try {
-    const accessToken = req.headers['authorization'];
-    const result = await jwt.verify(accessToken, secret);
+    const foundedUser = await User.findOne({
+      _id: req.accessTokenData.payload,
+    });
+    if (!foundedUser.permission.news.D) {
+      throw new Error(`Нет прав на удаление новостей!`);
+    }
 
     const foundedNews = await News.deleteOne({ _id: req.params.id });
     if (!foundedNews) {
@@ -114,13 +118,12 @@ module.exports.deleteNews = async (req, res) => {
 
     const foundedAllNews = await News.find();
     if (!foundedAllNews) {
-      throw new Error(`Новостей нет!`);
+      res.json([]);
     }
 
-    const preparedAllNews = [];
-    foundedAllNews.map((news) =>
-      preparedAllNews.push(serialize.serializeNews(news))
-    );
+    const preparedAllNews = foundedAllNews.map((news) => {
+      return serialize.serializeNews(news);
+    });
 
     res.json(preparedAllNews);
   } catch (err) {
